# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0", 
        "title": "Bad Bank App",
        "description": "Portfolio project for MIT Backend Course"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "tags": [
        {
            "name": "Auth",
            "description": "Authentication routes"
        },
        {
            "name": "Bank Function",
            "description": "Bank functionality routes"
        }
    ],
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/account/create/{name}/{email}/{password}": {
            "get": {
                "tags": ["Auth"],
                "description": "Create a new user",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Name of user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "path",
                        "description": "Email of user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "path",
                        "description": "Password of user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Get new user",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                }
            }
        },
        "/account/login/{email}/{password}": {
            "get": {
                "tags": ["Auth"],
                "description": "Login to the application",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "description": "Email of user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "path",
                        "description": "Password of user",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Get user info",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                }
            }
        },
        "/account/find/{email}": {
            "get": {
                "tags": ["Auth"],
                "description": "Find users with the email",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "description": "Email of user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Get user info",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/User"
                            }
                        }
                    }
                }
            }
        },
        "/account/findOne/{email}": {
            "get": {
                "tags": ["Auth"],
                "description": "Find a single user with the email",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "description": "Email of user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Get user info",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                }
            }
        },
        "/account/all": {
            "get": {
                "tags": ["Auth"],
                "description": "Find all users",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Get all user info",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/User"
                            }
                        }
                    }
                }
            }
        },
        "/account/update/{email}/{amount}": {
            "get": {
                "tags": ["Bank Function"],
                "description": "Deposit and withdraw money",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "description": "Email of user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "amount",
                        "in": "path",
                        "required": true,
                        "description": "Amount to be withdrawn/deposited",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Get updated user info",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "User": {
            "required": ["_id","name", "email", "balance"],
            "properties": {
                "_id": {
                    "type": "integer",
                    "uniqueItems": true
                },
                "name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "balance": {
                    "type": "integer"
                }
            }
        }
    }
}